Check docker-compose version.

+++ before_install.sh
@@ -1,3 +1,10 @@
+# Check docker-compose version. {{{1
+v=`docker-compose version | grep 'docker-compose version'`
+v=${v:23}
+v=${v%%,*}
+[ v = "$1" ] && exit
+
+# Install another version of docker-compose. {{{1
 sudo rm /usr/local/bin/docker-compose
 url_prefix=https://github.com/docker/compose/releases/download
 url_suffix=/$1/docker-compose-`uname -s`-`uname -m`
